몬스터, 플레이어, 자료형 수정
몬스터, 플레이어, 자료형 수정

+++ Server/Server/Monsterpool.h
@@ -11,6 +11,6 @@ public:
 	virtual ~CMonsterpool();
 
 public:
-	static concurrent_unordered_map<unsigned int, CMonster*> m_cumMonsterPool;
+	static concurrent_unordered_map<unsigned short, CMonster*> m_cumMonsterPool;
 };
